Implemented Save and Save As features

+++ 10 Projects/4) TextEditor.py
@@ -1,5 +1,6 @@
 import PySimpleGUI as sg
 from pathlib import Path
+import os
 
 smileys = [
         "happy",[":)",":D","<3"],
@@ -11,7 +12,7 @@ smiley_events = smileys[1] + smileys[3] + smileys[5]
 def makeWindow():
     sg.theme("DarkGreen4")
     menuLayout = [
-        ["File",["Open", "Open New Window","Save","---","Exit"]],
+        ["File",["Open", "Open New Window","Save As","Save","---","Exit"]],
         ["Tools",["Word Count"]],
         ["Add",smileys],
     ]
@@ -23,6 +24,12 @@ def makeWindow():
     ]
     return sg.Window("Text Editor", layout)
 
+def saveAs():
+    filePath = sg.popup_get_file("Save as", no_window=True, keep_on_top=True, modal=True, save_as=True) + ".txt"
+    file = Path(filePath)
+    file.write_text(values["-TEXT BOX-"])
+    window["-DOC NAME-"].update(Path(filePath).stem)
+
 window = makeWindow()
 
 while True:
@@ -31,16 +38,24 @@ while True:
         break
     
     if event == "Open":
-        filePath = sg.popup_get_file("open", no_window=True)
+        filePath = sg.popup_get_file("Open", no_window=True)
         if filePath:
+            savePath = filePath
             file = Path(filePath)
             window["-TEXT BOX-"].update(file.read_text())
-            window["-DOC NAME-"].update(filePath.split("/")[-1])
+            window["-DOC NAME-"].update(Path(filePath).stem)
+    
+    if event == "Save As":
+        saveAs()
+
     if event == "Save":
-        filePath = sg.popup_get_file("Save as", no_window=True, save_as=True) + ".txt"
-        file = Path(filePath)
-        file.write_text(values["-TEXT BOX-"])
-        window["-DOC NAME-"].update(filePath.split("/")[-1])
+        if os.path.exists(savePath):
+            file = Path(filePath)
+            file.write_text(values["-TEXT BOX-"])
+            window["-DOC NAME-"].update(Path(filePath).stem)
+        else: 
+            saveAs()
+
     if event == "Exit":
         break
     if event == "Open New Window":
